refactor: 로직 수정
