+++ models/repositories/user.verify.repository.ts
@@ -1,7 +1,5 @@
 import UserVerify from 'models/entities/user.verify';
 import { Repository, EntityRepository } from 'typeorm';
-import { v4 as uuid } from 'uuid';
-import * as bcrypt from 'bcrypt';
 
 /**
  * 기존에 사용자 아이디가 있으면 새로 레코드를 만들지 않고 토큰을 업데이트
@@ -11,14 +9,9 @@ import * as bcrypt from 'bcrypt';
  */
 @EntityRepository(UserVerify)
 class UserVerifyRepository extends Repository<UserVerify> {
-  public async createOne(email: string) {
-    const rand = uuid();
-    const verifyToken = `${email}|${rand}`;
-    const hashed = await bcrypt.hash(verifyToken, 12);
-
-    const flag = await this.findOne({ email });
-    const userVerify = !flag ? new UserVerify() : flag;
-    userVerify.token = hashed;
+  public async createOne(email: string, verifyToken: string) {
+    const userVerify = new UserVerify();
+    userVerify.token = verifyToken;
     userVerify.expiredAt = new Date(Date.now() + 1000 * 60 * 30);
     userVerify.email = email;
     await userVerify.save();
