--- services/app.service.ts
@@ -1,10 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-class AppService {
-  getHello(): string {
-    return 'Hello World!';
-  }
-}
-
-export default AppService;
